Knight500.ino
# Knight5000 project
The second part of this software cycle. Added hacking tools ESSID Spamer is a basic funny joking bomber demo for the LOL’z. The second version is ESSID  jammer  - this one is brutal and uses quad attacking via random is digits and  is Alpha numeric tricks also included is the ESSID name Spam attacking of the 1st demo. Tracking via channel and packets sent also displayed as real time stats.